llm-hallucination-firewall/services/decision_engine/requirements.txt
pydantic==2.5.0
pyyaml==6.0.1
